Fix MAC handling for DHCPv6
Always send MAC as a DUID, otherwise not hwaddr is send to DHCP. The
DUID will be some autogenerated one, making it more difficult to
identify machines, when providing lease data. Without MAC as DUID, no
MAC-based client classes can be created in the DHCP configuration which
is very well-desired functionality as well.

S.
https://www.freedesktop.org/software/systemd/man/networkd.conf.html#DUIDType=1
for details